requirements.txt
asgiref==3.10.0
awscli==1.43.15
black==25.9.0
botocore==1.42.9
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
cloudinary==1.44.1
colorama==0.4.6
dj-database-url==3.0.1
dj-rest-auth==7.0.1
Django==5.2.7
django-allauth==65.12.1
django-anymail==13.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
docutils==0.19
drf-yasg==1.21.11
environs==14.5.0
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
jmespath==1.0.1
marshmallow==4.1.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytokens==0.2.0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rsa==4.7.2
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.11.0
